style: :lipstick: Reorder imports components, comments unnecessary and no use variables
