+++ src/commons/utils/dateUtils.js
@@ -4,10 +4,5 @@ export const isDatePassed = (dateString) => {
   const currentDate = DateTime.local();
   const paramDate = DateTime.fromISO(dateString);
 
-  console.log(
-    '***currentDate, paramDate***',
-    currentDate.toString(),
-    paramDate.toString(),
-  );
   return currentDate >= paramDate;
 };
