+++ src/components/Header/Header.jsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { useDispatch } from 'react-redux';
+
 import {
   Navbar,
-  MobileNav,
   IconButton,
   Typography,
   Collapse,
