+++ src/components/Header/ProfileMenu/utils.js
@@ -1,0 +1,19 @@
+const profileMenuItems = [
+  {
+    label: 'Edit Profile',
+    icon: Cog6ToothIcon,
+    link: PROFILE,
+  },
+  {
+    label: 'Map',
+    icon: InboxArrowDownIcon,
+    link: MAP,
+  },
+  {
+    label: 'Sign Out',
+    icon: PowerIcon,
+    link: LOGIN,
+  },
+];
+
+export { profileMenuItems };
