+++ src/components/HelperMap/HelperMap.jsx
@@ -1,4 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react';
+
 import { Chip } from '@material-tailwind/react';
 import { PaperAirplaneIcon } from '@heroicons/react/24/outline';
