+++ src/components/MapComponent/MapComponent.css
@@ -1,5 +1,4 @@
 .map-container-component {
   height: 100%;
   width: 100%;
-  /* padding: 10px; */
 }
