+++ src/components/MapComponent/MarkerComponent/ChipAccessibility/ChipAccessibility.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import PropTypes from 'prop-types';
 import { CheckIcon } from '@heroicons/react/24/outline';
 import { Chip } from '@material-tailwind/react';
