+++ src/components/PlacePhoto/PlacePhoto.jsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+
 import PropTypes from 'prop-types';
 
 const PlacePhoto = ({ photoReference, children }) => {
