+++ src/hooks/useRefreshToken.js
@@ -1,14 +1,14 @@
 import { useEffect, useState } from 'react';
+
 import { useDispatch, useSelector } from 'react-redux';
+
 import {
-  logout,
   selectRefreshToken,
   selectTokenExpiredDate,
   selectUserData,
-  setIsTokenExpired,
 } from '../store/slices/authSlice';
 import { isDatePassed } from '../commons/utils/dateUtils';
-import { login, refreshTokenAction } from '../store/actions/authActions';
+import { refreshTokenAction } from '../store/actions/authActions';
 
 const useRefreshToken = () => {
   const dispatch = useDispatch();
@@ -18,16 +18,12 @@ const useRefreshToken = () => {
   const [isRefreshTokenDone, setIsRefreshTokenDone] = useState(true);
 
   useEffect(() => {
-    console.log('tokenExpiredDate', tokenExpiredDate);
     if (tokenExpiredDate && userData) {
-      console.log('estoy dentro de: tokenExpiredDate');
-
       const minutes = 33;
       const interval = minutes * 60 * 1000;
       const intervalRefreshToken = setInterval(async () => {
         const isExpired = isDatePassed(tokenExpiredDate);
         if (isExpired) {
-          console.log('está expirado***');
           try {
             const resultAction = await dispatch(
               refreshTokenAction({
@@ -36,14 +32,11 @@ const useRefreshToken = () => {
             );
 
             if (resultAction?.payload?.token) {
-              console.log('HA IDO BIEN EL REFRESH!!');
               setIsRefreshTokenDone(true);
             } else {
-              console.log('HA IDO MALLLLLL EL REFRESH!!');
               setIsRefreshTokenDone(false);
             }
           } catch (error) {
-            console.log('refresh token error: ', error);
             setIsRefreshTokenDone(false);
           }
         }
