+++ src/pages/Contact/Contact.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { ContactForm } from '../../components/ContactForm/ContactForm';
 
 import './Contact.css';
