+++ src/pages/Layout/Layout.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { Outlet } from 'react-router-dom';
 import { Header } from '../../components/Header/Header';
 import { Footer } from '../../components/Footer/Footer';
