+++ src/pages/NotFound/NotFound.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
+
 import { Link } from 'react-router-dom';
 
-import './NotFound.css';
 import { HOME } from '../../config/routes';
+
+import './NotFound.css';
 
 const NotFound = () => {
   return (
