+++ src/pages/Profile/Profile.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import { ProfileEditForm } from '../../components/ProfileEditForm/ProfileEditForm.jsx';
 
 import './Profile.css';
